+++ changelog.md
@@ -1,3 +1,5 @@
 * Updated to 1.21.11
 * Added Vibranium, Adamantium & Nether Spears
+* Added Vibranium & Adamantium Horse Armors
+* Added Vibranium & Adamantium Nautilus Armors
 * Migrated code to Mojang mappings